Rename CompletionCTA → GenerationSuccessBanner for clarity
